Removing log.fatalln usage
Removed Instances of log.fataln after recommendation by @nbkrause. Entire app shouldn't exit if there is an error with the tracing. Should just log the error and continue serving quotes. Any error in buildTracer() will get returned and picked up by the check on line 386, so the app will just not use the middle-ware for tracing if there is a problem with it's configuration instead of exiting.